Add tests and split the files for them
