+++ src/easings.js
@@ -1,0 +1,6 @@
+// Easings
+
+export function parseEasingParameters(string) {
+  const match = /\(([^)]+)\)/.exec(string);
+  return match ? match[1].split(',').map(p => parseFloat(p)) : [];
+}
