+++ src/index.js
@@ -1,357 +1,32 @@
-// Defaults
 
-const defaultInstanceSettings = {
-  update: null,
-  begin: null,
-  loopBegin: null,
-  changeBegin: null,
-  change: null,
-  changeComplete: null,
-  loopComplete: null,
-  complete: null,
-  loop: 1,
-  direction: 'normal',
-  autoplay: true,
-  timelineOffset: 0
-}
+import {
+  defaultInstanceSettings,
+  defaultTweenSettings,
+  validTransforms,
+  validSvgAttributes
+} from './constants';
 
-const defaultTweenSettings = {
-  duration: 1000,
-  delay: 0,
-  endDelay: 0,
-  easing: 'easeOutElastic(1, .5)',
-  round: 0
-}
+import {
+  minMax,
+  stringContains,
+  applyArguments,
+  is
+} from './utils';
 
-const validTransforms = ['translateX', 'translateY', 'translateZ', 'rotate', 'rotateX', 'rotateY', 'rotateZ', 'scale', 'scaleX', 'scaleY', 'scaleZ', 'skew', 'skewX', 'skewY', 'perspective'];
+import {
+  rgbToRgba,
+  hexToRgba,
+  hslToRgba,
+  colorToRgb
+} from './colors';
 
-const validSvgAttributes = [
-  'accent-height',
-  'accumulate',
-  'additive',
-  'alignment-baseline',
-  'allowReorder',
-  'alphabetic',
-  'amplitude',
-  'arabic-form',
-  'ascent',
-  'attributeName',
-  'attributeType',
-  'autoReverse',
-  'azimuth',
-  'BSection',
-  'baseFrequency',
-  'baseline-shift',
-  'baseProfile',
-  'bbox',
-  'begin',
-  'bias',
-  'by',
-  'CSection',
-  'calcMode',
-  'cap-height',
-  'class',
-  'clip',
-  'clipPathUnits',
-  'clip-path',
-  'clip-rule',
-  'color',
-  'color-interpolation',
-  'color-interpolation-filters',
-  'color-profile',
-  'color-rendering',
-  'contentScriptType',
-  'contentStyleType',
-  'cursor',
-  'cx',
-  'cy',
-  'DSection',
-  'd',
-  'decelerate',
-  'descent',
-  'diffuseConstant',
-  'direction',
-  'display',
-  'divisor',
-  'dominant-baseline',
-  'dur',
-  'dx',
-  'dy',
-  'ESection',
-  'edgeMode',
-  'elevation',
-  'enable-background',
-  'end',
-  'exponent',
-  'externalResourcesRequired',
-  'FSection',
-  'fill',
-  'fill-opacity',
-  'fill-rule',
-  'filter',
-  'filterRes',
-  'filterUnits',
-  'flood-color',
-  'flood-opacity',
-  'font-family',
-  'font-size',
-  'font-size-adjust',
-  'font-stretch',
-  'font-style',
-  'font-variant',
-  'font-weight',
-  'format',
-  'from',
-  'fr',
-  'fx',
-  'fy',
-  'GSection',
-  'g1',
-  'g2',
-  'glyph-name',
-  'glyph-orientation-horizontal',
-  'glyph-orientation-vertical',
-  'glyphRef',
-  'gradientTransform',
-  'gradientUnits',
-  'HSection',
-  'hanging',
-  'height',
-  'href',
-  'hreflang',
-  'horiz-adv-x',
-  'horiz-origin-x',
-  'ISection',
-  'id',
-  'ideographic',
-  'image-rendering',
-  'in',
-  'in2',
-  'intercept',
-  'KSection',
-  'k',
-  'k1',
-  'k2',
-  'k3',
-  'k4',
-  'kernelMatrix',
-  'kernelUnitLength',
-  'kerning',
-  'keyPoints',
-  'keySplines',
-  'keyTimes',
-  'LSection',
-  'lang',
-  'lengthAdjust',
-  'letter-spacing',
-  'lighting-color',
-  'limitingConeAngle',
-  'local',
-  'MSection',
-  'marker-end',
-  'marker-mid',
-  'marker-start',
-  'markerHeight',
-  'markerUnits',
-  'markerWidth',
-  'mask',
-  'maskContentUnits',
-  'maskUnits',
-  'mathematical',
-  'max',
-  'media',
-  'method',
-  'min',
-  'mode',
-  'NSection',
-  'name',
-  'numOctaves',
-  'OSection',
-  'offset',
-  'opacity',
-  'operator',
-  'order',
-  'orient',
-  'orientation',
-  'origin',
-  'overflow',
-  'overline-position',
-  'overline-thickness',
-  'PSection',
-  'panose-1',
-  'paint-order',
-  'path',
-  'pathLength',
-  'patternContentUnits',
-  'patternTransform',
-  'patternUnits',
-  'ping',
-  'pointer-events',
-  'points',
-  'pointsAtX',
-  'pointsAtY',
-  'pointsAtZ',
-  'preserveAlpha',
-  'preserveAspectRatio',
-  'primitiveUnits',
-  'RSection',
-  'r',
-  'radius',
-  'referrerPolicy',
-  'refX',
-  'refY',
-  'rel',
-  'rendering-intent',
-  'repeatCount',
-  'repeatDur',
-  'requiredExtensions',
-  'requiredFeatures',
-  'restart',
-  'result',
-  'rotate',
-  'rx',
-  'ry',
-  'SSection',
-  'scale',
-  'seed',
-  'shape-rendering',
-  'slope',
-  'spacing',
-  'specularConstant',
-  'specularExponent',
-  'speed',
-  'spreadMethod',
-  'startOffset',
-  'stdDeviation',
-  'stemh',
-  'stemv',
-  'stitchTiles',
-  'stop-color',
-  'stop-opacity',
-  'strikethrough-position',
-  'strikethrough-thickness',
-  'string',
-  'stroke',
-  'stroke-dasharray',
-  'stroke-dashoffset',
-  'stroke-linecap',
-  'stroke-linejoin',
-  'stroke-miterlimit',
-  'stroke-opacity',
-  'stroke-width',
-  'style',
-  'surfaceScale',
-  'systemLanguage',
-  'TSection',
-  'tabindex',
-  'tableValues',
-  'target',
-  'targetX',
-  'targetY',
-  'text-anchor',
-  'text-decoration',
-  'text-rendering',
-  'textLength',
-  'to',
-  'transform',
-  'type',
-  'USection',
-  'u1',
-  'u2',
-  'underline-position',
-  'underline-thickness',
-  'unicode',
-  'unicode-bidi',
-  'unicode-range',
-  'units-per-em',
-  'VSection',
-  'v-alphabetic',
-  'v-hanging',
-  'v-ideographic',
-  'v-mathematical',
-  'values',
-  'vector-effect',
-  'version',
-  'vert-adv-y',
-  'vert-origin-x',
-  'vert-origin-y',
-  'viewBox',
-  'viewTarget',
-  'visibility',
-  'WSection',
-  'width',
-  'widths',
-  'word-spacing',
-  'writing-mode',
-  'XSection',
-  'x',
-  'x-height',
-  'x1',
-  'x2',
-  'xChannelSelector',
-  'xlink:actuate',
-  'xlink:arcrole',
-  'xlink:href',
-  'xlink:role',
-  'xlink:show',
-  'xlink:title',
-  'xlink:type',
-  'xml:base',
-  'xml:lang',
-  'xml:space',
-  'YSection',
-  'y',
-  'y1',
-  'y2',
-  'yChannelSelector',
-  'ZSection',
-  'z',
-  'zoomAndPan'
-];
+import { parseEasingParameters } from './easings';
 
 // Caching
 
 const cache = {
   CSS: {},
   springs: {}
-}
-
-// Utils
-
-function minMax(val, min, max) {
-  return Math.min(Math.max(val, min), max);
-}
-
-function stringContains(str, text) {
-  return str.indexOf(text) > -1;
-}
-
-function applyArguments(func, args) {
-  return func.apply(null, args);
-}
-
-const is = {
-  arr: a => Array.isArray(a),
-  obj: a => stringContains(Object.prototype.toString.call(a), 'Object'),
-  pth: a => is.obj(a) && a.hasOwnProperty('totalLength'),
-  svg: a => a instanceof SVGElement,
-  inp: a => a instanceof HTMLInputElement,
-  dom: a => a.nodeType || is.svg(a),
-  str: a => typeof a === 'string',
-  fnc: a => typeof a === 'function',
-  und: a => typeof a === 'undefined',
-  hex: a => /(^#[0-9A-F]{6}$)|(^#[0-9A-F]{3}$)/i.test(a),
-  rgb: a => /^rgb/.test(a),
-  hsl: a => /^hsl/.test(a),
-  col: a => (is.hex(a) || is.rgb(a) || is.hsl(a)),
-  key: a => !defaultInstanceSettings.hasOwnProperty(a) && !defaultTweenSettings.hasOwnProperty(a) && a !== 'targets' && a !== 'keyframes'
-}
-
-// Easings
-
-function parseEasingParameters(string) {
-  const match = /\(([^)]+)\)/.exec(string);
-  return match ? match[1].split(',').map(p => parseFloat(p)) : [];
 }
 
 // Spring solver inspired by Webkit Copyright © 2016 Apple Inc. All rights reserved. https://webkit.org/demos/spring/spring.js
@@ -634,56 +309,6 @@ function mergeObjects(o1, o2) {
   const o = cloneObject(o1);
   for (let p in o2) o[p] = is.und(o1[p]) ? o2[p] : o1[p];
   return o;
-}
-
-// Colors
-
-function rgbToRgba(rgbValue) {
-  const rgb = /rgb\((\d+,\s*[\d]+,\s*[\d]+)\)/g.exec(rgbValue);
-  return rgb ? `rgba(${rgb[1]},1)` : rgbValue;
-}
-
-function hexToRgba(hexValue) {
-  const rgx = /^#?([a-f\d])([a-f\d])([a-f\d])$/i;
-  const hex = hexValue.replace(rgx, (m, r, g, b) => r + r + g + g + b + b );
-  const rgb = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  const r = parseInt(rgb[1], 16);
-  const g = parseInt(rgb[2], 16);
-  const b = parseInt(rgb[3], 16);
-  return `rgba(${r},${g},${b},1)`;
-}
-
-function hslToRgba(hslValue) {
-  const hsl = /hsl\((\d+),\s*([\d.]+)%,\s*([\d.]+)%\)/g.exec(hslValue) || /hsla\((\d+),\s*([\d.]+)%,\s*([\d.]+)%,\s*([\d.]+)\)/g.exec(hslValue);
-  const h = parseInt(hsl[1], 10) / 360;
-  const s = parseInt(hsl[2], 10) / 100;
-  const l = parseInt(hsl[3], 10) / 100;
-  const a = hsl[4] || 1;
-  function hue2rgb(p, q, t) {
-    if (t < 0) t += 1;
-    if (t > 1) t -= 1;
-    if (t < 1/6) return p + (q - p) * 6 * t;
-    if (t < 1/2) return q;
-    if (t < 2/3) return p + (q - p) * (2/3 - t) * 6;
-    return p;
-  }
-  let r, g, b;
-  if (s == 0) {
-    r = g = b = l;
-  } else {
-    const q = l < 0.5 ? l * (1 + s) : l + s - l * s;
-    const p = 2 * l - q;
-    r = hue2rgb(p, q, h + 1/3);
-    g = hue2rgb(p, q, h);
-    b = hue2rgb(p, q, h - 1/3);
-  }
-  return `rgba(${r * 255},${g * 255},${b * 255},${a})`;
-}
-
-function colorToRgb(val) {
-  if (is.rgb(val)) return rgbToRgba(val);
-  if (is.hex(val)) return hexToRgba(val);
-  if (is.hsl(val)) return hslToRgba(val);
 }
 
 // Units
